feat(sql/storage): A cache is added to avoid check commands insertion
This cache is activated by a configuration variable called
'enable_command_cache' which can be 0 or 1 to represent a boolean.
In case, it is equal to 1, each time an event is received to write
to the database the check command, a check is done to verify this
command changed or not. It will be written only if it changed.